somewhat a stable contriversion
