+++ client/src/components/OrganizationTopRepos.tsx
@@ -1,0 +1,5 @@
+const OrganizationTopRepos = () => {
+  return <div>Organization Contribution Charts Not Available</div>;
+};
+
+export default OrganizationTopRepos;
